Improve search & responsiveness at load. Fix bug where legend would get stuck at top of screen.
Fixed a bug where the legend would remain at the very top of the screen even if the user scrolled to the very top. This would happen when the resize event was fired and the legend was at the top of the screen, absolutely positioned. The absolute positioning led to canvas.offsetTop being 0, hence causing the canvas to remain at the top of the screen even when the scroll height was 1 (anything greater than 0).

Improved performance of search and responsiveness by changing NewList so that it draws the first 30 rivers at once, then draws the remaining ones in the background (rate varies based on device speed). This prevents search from lagging terribly on slower devices, improves initial page load responsiveness, and prepares the site for more rivers.