Fix quadratic memory usage in DataSource.js
Is appears that Promise.all creates a promise for every promise in the array. Either way, the ~2000 batches of individual Canadian gauges resulted in almost 2 million promises in memory, causisng massive issues.

This commit moved peak memory usage from ~600MB, to ~140MB in my testing, much of which is made up of conversion tables, etc.